Move registry related files to new package
