+++ src/main/java/de/tum/www1/orion/bridge/ArtemisJSBridge.java
@@ -24,9 +24,13 @@ import de.tum.www1.orion.dto.ProgrammingExercise;
 import de.tum.www1.orion.dto.RepositoryType;
 import de.tum.www1.orion.enumeration.ExerciseView;
 import de.tum.www1.orion.ui.util.ImportPathChooser;
-import de.tum.www1.orion.util.*;
+import de.tum.www1.orion.util.JsonUtilsKt;
+import de.tum.www1.orion.util.OrionProjectUtil;
+import de.tum.www1.orion.util.UtilsKt;
 import de.tum.www1.orion.util.project.OrionJavaInstructorProjectCreator;
 import de.tum.www1.orion.util.registry.OrionGlobalExerciseRegistryService;
+import de.tum.www1.orion.util.registry.OrionInstructorExerciseRegistry;
+import de.tum.www1.orion.util.registry.OrionStudentExerciseRegistry;
 import de.tum.www1.orion.vcs.CredentialsService;
 import de.tum.www1.orion.vcs.OrionGitUtil;
 import javafx.application.Platform;
